Attempting to make it discoverable easier
Updated the description.xml to contain the Modelname: Philips hue bridge 2012 (omitted the Router).  Also removed the tab indents and carriage returns (\r) because examples found on the web didn't seem to have these.

Also added in the GET /api/*/config option so it will return the JSON configuration description.  Some devices might be using this to distinguish the Hue.

Lastly, I added in the "ssdp:all" so the server responds to it as upnp:rootdevice.

I have no idea if it will help or not but it is the best I can find based on examples scattered over the net.